xhr/sync-xhr-supported-by-feature-policy.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('sync-xhr', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise sync-xhr.');
</script>
